+++ packages/metalsmith-html-sri/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.2.0 / 2019-12-29
+
+- Added `ignoreResources` option.
+- Removed ignoring `fonts.googleapis.com` by default.
+- Fixed overwriting the existing `integrity` attribute of remote resources.
+- Added `'use strict';`.
+
 ## v0.1.0 / 2019-05-25
 
 - Added Jest tests.